backend/pyproject.toml
[project]
name = "zenail-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "SQLAlchemy",
  "alembic",
  "psycopg[binary]",
  "python-jose[cryptography]",
  "passlib",
  "pydantic",
  "pydantic-settings",
  "email-validator",
  "python-multipart",
]

[dependency-groups]
dev = [
  "pytest",
  "httpx",
]

